refactor(voice-bot): restructure example to match standard pattern
- Rename demo-wav.ts to main.ts as primary entry point
- Remove unnecessary config.ts file
- Update package.json scripts to use main.ts entry
- Update README.md with current implementation details
- Align structure with other example projects
