+++ examples/voice-bot/tsconfig.json
@@ -6,7 +6,7 @@
     "composite": true,
     "jsx": "react-jsx"
   },
-  "include": ["src/**/*", "config.ts"],
+  "include": ["src/**/*"],
   "exclude": ["**/*.spec.ts", "**/*.test.ts", "dist", "node_modules"],
   "references": [{ "path": "../../packages/sdk" }]
 }
